+++ CHANGELOG.md
@@ -8,6 +8,46 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+## [0.11.0] - 2024-11-27
+
+### Added
+
+- Add metric package as middleware `go-chi/chi`, support `request_duration_millis`, `requests_inflight`, and `response_size_bytes` metric. ([#69])
+
+### Changed
+
+- Upgrade `go.opentelemetry.io/otel`, `go.opentelemetry.io/otel/sdk`, & `go.opentelemetry.io/otel/trace` to `v1.32.0`. ([#74])
+
+## [0.10.1] - 2024-10-27
+
+### Changed
+
+- Upgrade `go.opentelemetry.io/otel`, `go.opentelemetry.io/otel/sdk`, & `go.opentelemetry.io/otel/trace` to `v1.31.0`. ([#70])
+
+### Fixed
+
+- Span for websocket connection now won't be marked as error span. ([#67])
+
+## [0.10.0] - 2024-09-17
+
+### Added
+
+- Add `WithTraceResponseHeaders` option to include trace information in response headers, this option replaces the deprecated `WithTraceIDResponseHeader` option. ([#62])
+
+### Changed
+
+- Upgrade `go.opentelemetry.io/otel`, `go.opentelemetry.io/otel/sdk`, & `go.opentelemetry.io/otel/trace` to `v1.30.0`. ([#64])
+- Set the go versions for testing in both `Makefile` & `compatibility-test.yml` to `1.22` & `1.23`. ([#64])
+
+### Fixed
+
+- Fix Gobrew installation in CI pipeline. ([#63])
+
+### Removed
+
+- Drop support for Go 1.21. ([#64])
+- Deprecated `WithTraceIDResponseHeader` option, use `WithTraceResponseHeaders` instead. ([#62])
+
 ## [0.9.0] - 2024-07-06
 
 ### Changed
@@ -174,6 +214,13 @@ It contains instrumentation for trace and depends on:
 - Example code for a basic usage.
 - Apache-2.0 license.
 
+[#74]: https://github.com/riandyrn/otelchi/pull/74
+[#70]: https://github.com/riandyrn/otelchi/pull/70
+[#69]: https://github.com/riandyrn/otelchi/pull/69
+[#67]: https://github.com/riandyrn/otelchi/pull/67
+[#64]: https://github.com/riandyrn/otelchi/pull/64
+[#63]: https://github.com/riandyrn/otelchi/pull/63
+[#62]: https://github.com/riandyrn/otelchi/pull/62
 [#49]: https://github.com/riandyrn/otelchi/pull/49
 [#47]: https://github.com/riandyrn/otelchi/pull/47
 [#43]: https://github.com/riandyrn/otelchi/pull/43
@@ -196,7 +243,10 @@ It contains instrumentation for trace and depends on:
 [#2]: https://github.com/riandyrn/otelchi/pull/2
 [#1]: https://github.com/riandyrn/otelchi/pull/1
 
-[Unreleased]: https://github.com/riandyrn/otelchi/compare/v0.9.0...HEAD
+[Unreleased]: https://github.com/riandyrn/otelchi/compare/v0.11.0...HEAD
+[0.11.0]: https://github.com/riandyrn/otelchi/releases/tag/v0.11.0
+[0.10.1]: https://github.com/riandyrn/otelchi/releases/tag/v0.10.1
+[0.10.0]: https://github.com/riandyrn/otelchi/releases/tag/v0.10.0
 [0.9.0]: https://github.com/riandyrn/otelchi/releases/tag/v0.9.0
 [0.8.0]: https://github.com/riandyrn/otelchi/releases/tag/v0.8.0
 [0.7.0]: https://github.com/riandyrn/otelchi/releases/tag/v0.7.0